fix: all warnings on MSVC

+++ source/Transform/Dispatch/Generic/TransformAffineTiledMT_Bilinear_MEM_Generic.ipp
@@ -39,8 +39,8 @@ InvokeTransformAffineTiledMTProcessScanline_Bilinear_MEM_Generic( tByte* iDst, i
     const int maxx = minx + info.src_roi.w;
     const int maxy = miny + info.src_roi.h;
     for( int x = 0; x < info.dst_roi.w; ++x ) {
-        const float modx = FMaths::PyFModulo( point_in_src.x, info.src_roi.w );
-        const float mody = FMaths::PyFModulo( point_in_src.y, info.src_roi.h );
+        const float modx = FMaths::PyFModulo( point_in_src.x, static_cast< float >( info.src_roi.w ) );
+        const float mody = FMaths::PyFModulo( point_in_src.y, static_cast< float >( info.src_roi.h ) );
         const int   left    = static_cast< int >( modx );
         const int   top     = static_cast< int >( mody );
         const int   right   = ( left + 1 ) % info.src_roi.w;
